+++ mobile/android/android-components/docs/changelog.md
@@ -12,6 +12,9 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
 
+* **feature-remotetabs**
+  * Add new `RemoteTabsFeature` to view tabs from other synced devices and upload our own.
+
 # 24.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v23.0.0...v24.0.0)